Modelo de Pago BD

+++ database/migrations/2016_03_16_154842_create_heritage_property_table.php
@@ -16,7 +16,7 @@ class CreateHeritagePropertyTable extends Migration
             $table->increments('id');
             $table->string('document_code');
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price',30,15);
             $table->string('address');
             $table->timestamps();
         });
