+++ database/migrations/2016_03_16_160602_create_user_assets_table.php
@@ -15,12 +15,12 @@ class CreateUserAssetsTable extends Migration
         Schema::create('user_assets', function (Blueprint $table) {
             $table->increments('id')->unsigned();
             $table->integer('user_id')->unsigned();
-            $table->float('monthly_contributions');
+            $table->decimal('monthly_contributions',30,15);
             $table->string('reason');
             $table->string('payment');
             $table->string('payment_number');
-            $table->float('accumulated');
-            $table->float('bloked')->default(0.0);
+            $table->decimal('accumulated',30,15);
+            $table->decimal('bloked',30,15)->default(0.0);
 
             $table->foreign('user_id')
                 ->references('id')
